LockUpdate2KeyDisk: Part3 - wwid based devlock
Cleaned up gudbud1, a simple shorten logic.

Implemented a simple devlock mechanism. It derives a new key for
enc/decode by using the original key and the unique wwid of the
given scsi storage device (which could even be a usb storage device)
Inturn the original key is replaced by this newly derived key.

One can specify the disk that will be used for getting the wwid used
for locking update to itself (i.e the key disk).

If one specifies the update disk containing the update image itself
as the keydisk, then one can't duplicate that disk into another
disk and get the update to work.

If one specifies a different disk as the keydisk, then one can
duplicate the update disk as many times as one wants. But they
cant use either the original or the duplicated update disk, unless
they have the keydisk with them.