LockUpdate2KeyDisk: Linked KeyDiskModelPrefix cmdline arg to dd
The new commandline argument (which is the last in the command line)
which specifies the keyDisks model prefix, is used to identify the
linux device name of the key disk already in the find_sd logic.

Now this key disk name is inturn passed to the dd_s2d logic and
it inturn derives a new key for its enc/dec operations by calling
the procd1_devlock before doing any transfer operations.