User controls key update b4 next block processed
Added a new commandline argument nbMod (NextBlockModified).

This is inturn chained into procd1_e1_next throu dd.

With this now user has some amount of control over how the key used
for enc/dec is updated before it is used to process next block of
data.

Final control is as always with developer, as to how much control
is provided to user, by deciding how to use nbMod passed to their
_next logic.