Made some msgs less informative - hide intentions
Added a SAFE_MSGS alternative for some critical operations msgs
in the logic, so that the intention of the code is not directly
obvious from the message printed.

By default the prg prints are informative. However if one defines
SAFE_MSGS, while compiling, then some of the prints become less
informative.

Makefile updated to generate both safe and unsafe print based vers
of the program.