Added SAFE_ARGS logic and made it default
Now the arguments SrcOffset and DstOffset are no longer used directly,
instead they are processed before being used.

SrcOffset is substracted by 0x5a5a5a5a, so add 0x5a5a5a5a to actual
value before passing it to the program.

DstOffset is substracted by 0xa5a5a5a5, so add 0xa5a5a5a5 to actual
value before passing it to the program.

THis is so that if someone looks at the arguments being passed to the
program, they can't identify the locations used, unless they
disassemble the program and or step through it and or strace it and
look at the arguments finally used.